move the model-based builders to core.model.graphql package
