+++ aws-api/src/main/java/com/amplifyframework/core/model/graphql/ModelMutation.java
@@ -13,8 +13,9 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.api.aws;
+package com.amplifyframework.core.model.graphql;
 
+import com.amplifyframework.api.aws.AppSyncGraphQLRequestFactory;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.MutationType;
 import com.amplifyframework.core.model.Model;
