+++ rxbindings/src/test/java/com/amplifyframework/rx/RxApiBindingTest.java
@@ -22,9 +22,6 @@ import com.amplifyframework.api.ApiCategory;
 import com.amplifyframework.api.ApiCategoryConfiguration;
 import com.amplifyframework.api.ApiException;
 import com.amplifyframework.api.ApiPlugin;
-import com.amplifyframework.api.aws.ModelMutation;
-import com.amplifyframework.api.aws.ModelQuery;
-import com.amplifyframework.api.aws.ModelSubscription;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.GraphQLResponse;
 import com.amplifyframework.api.rest.RestOptions;
@@ -32,6 +29,9 @@ import com.amplifyframework.api.rest.RestResponse;
 import com.amplifyframework.core.Action;
 import com.amplifyframework.core.Consumer;
 import com.amplifyframework.core.model.Model;
+import com.amplifyframework.core.model.graphql.ModelMutation;
+import com.amplifyframework.core.model.graphql.ModelQuery;
+import com.amplifyframework.core.model.graphql.ModelSubscription;
 import com.amplifyframework.testutils.random.RandomModel;
 import com.amplifyframework.testutils.random.RandomString;
