+++ testutils/src/main/java/com/amplifyframework/testutils/sync/SynchronousApi.java
@@ -20,12 +20,11 @@ import androidx.annotation.NonNull;
 import com.amplifyframework.api.ApiCategory;
 import com.amplifyframework.api.ApiCategoryBehavior;
 import com.amplifyframework.api.ApiException;
-import com.amplifyframework.api.aws.ModelMutation;
-import com.amplifyframework.api.aws.ModelQuery;
-import com.amplifyframework.api.aws.ModelSubscription;
+import com.amplifyframework.core.model.graphql.ModelMutation;
+import com.amplifyframework.core.model.graphql.ModelQuery;
+import com.amplifyframework.core.model.graphql.ModelSubscription;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.GraphQLResponse;
-import com.amplifyframework.api.graphql.SubscriptionType;
 import com.amplifyframework.api.rest.RestOptions;
 import com.amplifyframework.api.rest.RestResponse;
 import com.amplifyframework.core.async.Cancelable;
